+++ Sources/Extensions/SwiftStdlib/SignedIntegerExtensions.swift
@@ -78,5 +78,21 @@ public extension SignedInteger {
 		return (self * n).abs / gcd(of: n)
 	}
 	// swiftlint:enable identifier_name
-
+    
+    #if canImport(Foundation)
+    @available(iOS 9.0, macOS 10.11, *)
+    /// SwifterSwift: Ordinal representation of an integer.
+    ///
+    ///        print((12).ordinalString()) // prints "12th"
+    ///
+    /// - Parameter locale: Locale, default is .current.
+    /// - Returns: String ordinal representation of number in specified locale language. E.g. input 92, output in "en": "92nd"
+    public func ordinalString(locale: Locale = .current) -> String? {
+            let formatter = NumberFormatter()
+            formatter.locale = locale
+            formatter.numberStyle = .ordinal
+            guard let number = self as? NSNumber else { return nil }
+            return formatter.string(from: number)
+    }
+    #endif
 }